+++ Geo/src/main/java/me/nettrash/geo/di/AppModule.kt
@@ -9,6 +9,7 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import me.nettrash.geo.data.db.GeoDatabase
 import me.nettrash.geo.data.db.HistoryDao
+import me.nettrash.geo.data.db.TripDao
 import javax.inject.Singleton
 
 @Module
@@ -22,11 +23,16 @@ object AppModule {
             context,
             GeoDatabase::class.java,
             "geo_database"
-        ).addMigrations(GeoDatabase.MIGRATION_1_2).build()
+        ).addMigrations(GeoDatabase.MIGRATION_1_2, GeoDatabase.MIGRATION_2_3).build()
     }
 
     @Provides
     fun provideHistoryDao(database: GeoDatabase): HistoryDao {
         return database.historyDao()
+    }
+
+    @Provides
+    fun provideTripDao(database: GeoDatabase): TripDao {
+        return database.tripDao()
     }
 }